added bounded_mlp_policy

+++ baselines/envs/lqg1d.py
@@ -2,6 +2,7 @@
 from numbers import Number
 
 import gym
+import gym.spaces
 from gym import spaces
 from gym.utils import seeding
 import numpy as np
@@ -59,8 +60,8 @@ class LQG1D(gym.Env):
         high = np.array([self.max_pos])
         self.action_space = spaces.Box(low=-self.max_action,
                                        high=self.max_action,
-                                       shape=(1,))
-        self.observation_space = spaces.Box(low=-high, high=high)
+                                       shape=(1,), dtype=np.float32)
+        self.observation_space = spaces.Box(low=-high, high=high,dtype=np.float32)
 
 
         # initialize state
